src/pegasus/agents/admin/requirements.txt
# Claude Agent Communication System - Administration Tools
# Requirements for comprehensive administration interface
# Supports CLI, Web Console, TUI, and deployment management

# Core dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
click>=8.1.7
rich>=13.7.0
pydantic>=2.5.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# Database and data handling
sqlite3  # Built-in with Python
sqlalchemy>=2.0.23
alembic>=1.13.0
redis>=5.0.1
aioredis>=2.0.1

# Monitoring and metrics
prometheus-client>=0.19.0
grafana-api>=1.0.3
requests>=2.31.0
aiohttp>=3.9.0
websockets>=12.0

# Container and orchestration
docker>=6.1.3
kubernetes>=28.1.0
pyyaml>=6.0.1

# Terminal and UI
urwid>=2.1.2
curses  # Built-in with Python
blessed>=1.20.0
textual>=0.45.0

# WebSocket and real-time communication
python-socketio>=5.10.0
socketio>=0.2.1
websocket-client>=1.6.4

# Security and authentication
cryptography>=41.0.8
bcrypt>=4.1.2
pyjwt>=2.8.0
python-ldap>=3.4.3

# System monitoring and metrics
psutil>=5.9.6
py-cpuinfo>=9.0.0
gpustat>=1.1.1
nvidia-ml-py3>=7.352.0

# Data processing and analysis
pandas>=2.1.4
numpy>=1.25.2
matplotlib>=3.8.2
seaborn>=0.13.0

# Configuration management
configparser  # Built-in with Python
toml>=0.10.2
python-dotenv>=1.0.0

# Async support
asyncio  # Built-in with Python
asyncpg>=0.29.0
aiosqlite>=0.19.0
aiodns>=3.1.1

# Testing (optional, for development)
pytest>=7.4.3
pytest-asyncio>=0.21.1
pytest-mock>=3.12.0
pytest-cov>=4.1.0

# Documentation (optional)
sphinx>=7.2.6
sphinx-rtd-theme>=1.3.0

# Development tools (optional)
black>=23.11.0
flake8>=6.1.0
mypy>=1.7.1
pre-commit>=3.6.0

# Performance profiling (optional)
py-spy>=0.3.14
memory-profiler>=0.61.0
line-profiler>=4.1.1

# Additional utilities
jinja2>=3.1.2
marshmallow>=3.20.1
cerberus>=1.3.5
schedule>=1.2.1
python-crontab>=3.0.0

# Network utilities
netaddr>=0.10.1
dnspython>=2.4.2
scapy>=2.5.0

# File and data format support
openpyxl>=3.1.2
xlsxwriter>=3.1.9
python-csv>=0.0.13

# Logging and debugging
structlog>=23.2.0
loguru>=0.7.2
sentry-sdk>=1.38.0

# Email notifications (optional)
emails>=0.6
smtplib  # Built-in with Python

# Backup and archiving
tarfile  # Built-in with Python
zipfile  # Built-in with Python
paramiko>=3.4.0  # SSH/SCP support

# Time and date handling
python-dateutil>=2.8.2
pytz>=2023.3
croniter>=2.0.1

# Image and visualization (optional)
pillow>=10.1.0
plotly>=5.17.0
bokeh>=3.3.2

# API clients
httpx>=0.25.2
requests-oauthlib>=1.3.1

# Development server tools
watchdog>=3.0.0
livereload>=2.6.3

# Deployment and infrastructure
ansible>=8.7.0
terraform-external-data>=2.3.3
boto3>=1.34.0  # AWS SDK
google-cloud-core>=2.4.1  # Google Cloud SDK
azure-mgmt-core>=1.4.0  # Azure SDK